requirements.txt
apache_beam
apache_beam[gcp]
numpy
pyspark
google-cloud-dataproc